update to include twl test
some systems, usually due to prior uninstallation of cfw, have a broken twl firm and cannot use frogminer. it is pointless for such users to proceed through these steps only to find they can't do it. this can be detected from stock very easily and this pr includes steps for such a test